move auth client adapter to module folder
