+++ src/app/(auth)/sign-in/components/SignInFormContainer.tsx
@@ -9,7 +9,7 @@ import { validateTextInput } from "@/utils/form/validateInput";
 import { clientSignIn } from "@/store/features/auth/authSlice";
 import { useAppDispatch } from "@/store/hooks";
 import routePaths from "@/utils/routePaths";
-import { type AuthClientAdapter } from "@/app/(auth)/_adapters/authClientAdapter";
+import { type AuthClientAdapter } from "@/modules/auth/adapters/primary/authClientAdapter";
 import { TYPES } from "@/di/types";
 import { resolve } from "@/di/resolver";
