+++ src/di/config.ts
@@ -2,7 +2,7 @@ import { container } from "tsyringe";
 import { TYPES } from "./types";
 import { AuthApiAdapter } from "@/modules/auth/adapters/secondary/authApiAdapter";
 import { AxiosAdapter } from "@/modules/restApi/adapters/secondary/AxiosAdapter";
-import { AuthClientAdapter } from "@/app/(auth)/_adapters/authClientAdapter";
+import { AuthClientAdapter } from "@/modules/auth/adapters/primary/authClientAdapter";
 import { LoginUsecase } from "@/modules/auth/application/usecases/loginUsecase";
 
 container.register(TYPES.RestApiPort, { useClass: AxiosAdapter });
